requirements.txt
pandas>=2.0.0
requests>=2.31.0
feedparser>=6.0.10
pytrends>=4.9.2
urllib3<2   # pytrends uses method_whitelist which was removed in urllib3 2.x
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
yfinance>=0.2.40
